requirements.txt
astroid==1.6.3
colorama==0.3.9
Django==2.0.3
django-filter==1.1.0
djangorestframework==3.7.7
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
mysqlclient==1.3.12
Pillow==5.1.0
pylint==1.8.4
PyMySQL==0.8.1
pytz==2018.3
Shapely==1.6.4.post1
six==1.11.0
wrapt==1.10.11
